refactor(user): 충돌 안나게 원래대로 돌리기
