+++ BuyOrNot/src/main/java/com/ambiguous/buyornot/mypage/MypageRepositoty.java
@@ -1,6 +1,6 @@
 package com.ambiguous.buyornot.mypage;
 
-import com.ambiguous.buyornot.user.api.domain.User;
+import com.ambiguous.buyornot.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
